company project workspace

+++ lib/features/projects/presentation/widgets/bloc_builder_mark_as_finish_button.dart
@@ -68,7 +68,7 @@ class BlocBuilderMarkAsFinishButton extends StatelessWidget {
                           PayProjectRequestModel(
                             projectId: projectModel.id,
                             freelancerId: projectModel.freelancerId,
-                            budget: projectModel.price,
+                            budget: projectModel.price.toInt(),
                             deliveryTime: projectModel.days.toString(),
                           ),
                         );
@@ -95,7 +95,7 @@ class BlocBuilderMarkAsFinishButton extends StatelessWidget {
                           PayProjectRequestModel(
                             projectId: projectModel.id,
                             freelancerId: projectModel.freelancerId,
-                            budget: projectModel.price,
+                            budget: projectModel.price.toInt(),
                             deliveryTime: projectModel.days.toString(),
                           ),
                         );
